Tao folder trong dialog upload, resize hinh anh khi upload

+++ app/Http/Controllers/Admin/UploadController.php
@@ -25,12 +25,20 @@ class UploadController extends BaseController
 
         $arr=array();
 
+        $now=time();
+
         foreach ($dir as $key) {
-            $time=filemtime($path."/".$key);
-            $size=filesize($path."/".$key);
-            list($width,$height)=getimagesize($path."/".$key);
-            $date=date("d/m/Y H:i",$time);
-            array_push($arr, array("name"=>$key,"time"=>$date,"size"=>$size,"width"=>$width,"height"=>$height));
+            
+            if(is_dir($path."/".$key))
+            {
+                 array_push($arr, array("name"=>$key,"time"=>'9999999999'));
+            }else{
+                $time=filemtime($path."/".$key);
+                $size=filesize($path."/".$key);
+                list($width,$height)=getimagesize($path."/".$key);
+                $date=date("d/m/Y H:i",$time);
+                array_push($arr, array("name"=>$key,"time"=>$date,"size"=>$size,"width"=>$width,"height"=>$height));
+            }
         }
 
         return json_encode($arr);
@@ -45,11 +53,68 @@ class UploadController extends BaseController
 
 	public function removeimg(){
 		$path=public_path()."/image/".$_POST['file'];
+        if(file_exists($path)){
+            $result=unlink($path);
 
-        $result=unlink($path);
-
-        return json_encode($result);
+            return json_encode($result);
+        }
+        return -1;
 	}
+
+    public function removefolder(){
+        $path=public_path()."/image/".$_POST['file'];
+        if(file_exists($path)){
+            $dir=scandir($path);
+
+            if(count($dir)==2)
+            {
+                if(rmdir($path)){
+                    return 1;
+                }
+                return -1;
+            }
+
+            return 3;
+        }
+        return 2;
+    }
+
+    public function createfolder(){
+        if(isset($_POST['foldername'])){
+            $path = public_path().'/image/'.$_POST['folderroot'].'/'.$_POST['foldername'];
+            if(!file_exists($path)){
+                if(mkdir($path))
+                return 1;
+             return -1;
+            }
+           
+            return -1;
+        }
+    }
+
+    public function loadonlyfolder(){
+        $path=public_path()."/image/".$_POST['folder'];
+        
+        if(!file_exists($path)){
+            return -1;
+        }
+
+        $dir=scandir($path);
+
+        unset($dir[0]);
+        unset($dir[1]);
+
+        $arr=array();
+
+        foreach ($dir as $key) {
+            if(is_dir($path."/".$key))
+            {
+                 $arr[]=$key;
+            }
+        }
+
+        return json_encode($arr);
+    }
 }
 
 ?>
